Moved EncoderProgressListener. Added a simple EchoingProgressListener to it wouldn’t be lonely in its new package.
