+++ jave-core-test/src/test/java/ws/schild/jave/PListener.java
@@ -9,6 +9,7 @@ import java.util.LinkedList;
 import java.util.List;
 
 import ws.schild.jave.info.MultimediaInfo;
+import ws.schild.jave.progress.EncoderProgressListener;
 
 /**
  *
